Move strategies to subdirs, add null + serialize pack.
