+++ src/Strategy/Compression/NullCompressionStrategy.php
@@ -1,5 +1,5 @@
 <?php
-namespace Aliance\Compressor\Strategy;
+namespace Aliance\Compressor\Strategy\Compression;
 
 class NullCompressionStrategy implements CompressionStrategyInterface
 {
