+++ src/Strategy/Pack/JsonPackStrategy.php
@@ -1,0 +1,23 @@
+<?php
+namespace Aliance\Compressor\Strategy\Pack;
+
+class JsonPackStrategy implements PackStrategyInterface
+{
+    /** @inheritdoc */
+    public function getMinLength()
+    {
+        return 150;
+    }
+
+    /** @inheritdoc */
+    public function pack($value)
+    {
+        return json_encode($value);
+    }
+
+    /** @inheritdoc */
+    public function unpack($value)
+    {
+        return json_decode($value, true);
+    }
+}
